data type corrections for measured voltages

+++ include/tDCDC.h
@@ -25,8 +25,8 @@ public:
 	double get_Ki();
 	double get_Kd();
 
-	uint16_t get_last_Vcurrent();
-	uint16_t get_last_Vnow();
+	double get_last_Vcurrent();
+	double get_last_Vnow();
 	uint16_t get_Vset();
 	uint16_t get_Vmax();
 	int16_t get_Verror_percent();
@@ -65,8 +65,8 @@ private:
 	double C0;
 	double C1;
 	double C2;
-	uint16_t last_Vnow;
-	uint16_t last_Vcurrent;
+	double last_Vnow;
+	double last_Vcurrent;
 
 	double get_HV_voltage(uint8_t nAvg); //deprecated
 	double get_HV_voltage_fast(float alpha);
